Updating interfaceMap along with rescanonebus
Description:
FruDevice service exit, when trying to map device from
busInterfaceMap on deviceMap.Sometimes, the mux device which is
detected in the rootbus on rescanBusees is not detected while
rescanOneBus.In this case the device which is not detected
is not updated in interfaceMap. Added condition to check and
erase the device from busInterfaceMap. Also chassis instance was
reduced after chassis power off/on. There is sdbus match for
"CurrentHostState" and handler checks the HostState endswith
"Running". "CurrentHostState" endswith "Running" on below two cases.
-TransitionToRunning
-Running
So the condition consider "TransitionToRunning" as "Running" and
rescanBusses initiated before host start running. In this case
reading of device failed and bus added into blacklist. So modified
condition to check "CurrentHostState" as
"xyz.openbmc_project.State.Host.HostState.Running"

Tested:
Successfully build and chassis instance count was not reduced.
	1. power on/off chassis and check chassis instance count
	   on redfish URI.
	   https://10.19.102.14/redfish/v1/Chassis
	{
    	"@odata.id": "/redfish/v1/Chassis",
    	"@odata.type": "#ChassisCollection.ChassisCollection",
    	"Members": [
        	{
            	"@odata.id": "/redfish/v1/Chassis/Cpld0"
        	},
        	{
            	"@odata.id": "/redfish/v1/Chassis/Cpld1"
        	},
        	{
            	"@odata.id": "/redfish/v1/Chassis/Luna_IOE0"
        	},
        	{
            	"@odata.id": "/redfish/v1/Chassis/Luna_IOE1"
        	},
        	{
            	"@odata.id": "/redfish/v1/Chassis/Luna_M_2_Riser"
        	},
        	{
            	"@odata.id": "/redfish/v1/Chassis/Luna_Midplane"
        	},
        	{
            	"@odata.id": "/redfish/v1/Chassis/Luna_Motherboard"
        	},
        	{
            	"@odata.id": "/redfish/v1/Chassis/Luna_PDB"
        	},
        	{
            	"@odata.id": "/redfish/v1/Chassis/Luna_Software"
        	}
    	],
    	"Members@odata.count": 9,
    	"Name": "Chassis Collection"
	}

Signed-off-by: selvaganapathi <selvaganapathim@ami.com>