fixed prunedProductName cannot be pruned Symptom: If there is more than 1 same FRU product name on the system, fru-device will append `_${num}` for it. For example, we will see the following strings if two cedar modules are installed on skinnyjoe. |-/xyz/openbmc_project/FruDevice/ConnectX_7_NDR_2P_NIC_____________________________ |-/xyz/openbmc_project/FruDevice/ConnectX_7_NDR_2P_NIC______________________________0
And the second string ends with 0 cannot be processed because we use find_last_not_of('_') to handle.
It will remain the same.

Change:
add a suffix pattern to replace the number with `_`. So, it can be processed.

tested and passed on the skinnyjoe

fixes jira https://jirasw.nvidia.com/browse/DGXOPENBMC-10407